backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
openai
google-generativeai
scikit-learn
streamlit
